refactor(flight): review follow-ups — single print, BaseException, e2e test joins the flight xdist group
Three review items on the abort-policy change plus a test move:

- streaming_split_exchange no longer wraps f in a second excepts_print_exc:
  with the reraise default the inner wrap only printed the same traceback
  twice. The unit test now pins exactly one Traceback in stderr.
- streaming_split_exchange uses exc=BaseException like the other two
  exchanges, so a SystemExit from a split's f is printed server-side too
  (abort behavior was already identical either way).
- FlightClient._do_action returns Iterator[Any] instead of the map builtin.
- The end-to-end abort test moves to flight/tests/test_flight_exchange.py:
  that conftest puts everything under it in the "flight" xdist group, which
  the test silently opted out of by living in common/utils/tests. The ~20
  lines of split helpers are duplicated rather than imported across test
  packages; docstrings in both files cross-reference the other half.
